[CRM-14] Add CarService interface, add pagination, code refactoring
