+++ backend/src/main/java/com/crm/dto/mapper/MapperConfiguration.java
@@ -1,11 +1,11 @@
-package com.crm.config;
+package com.crm.dto.mapper;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class AppConfiguration {
+public class MapperConfiguration {
 
     @Bean
     public ModelMapper modelMapper() {
