+++ backend/src/main/java/com/crm/service/CarService.java
@@ -1,27 +1,9 @@
 package com.crm.service;
 
-import com.crm.dto.mapper.CarMapper;
 import com.crm.dto.response.CarResponse;
-import com.crm.repository.CarRepository;
-import org.springframework.stereotype.Service;
+import org.springframework.data.domain.Page;
 
-import java.util.List;
-import java.util.stream.Collectors;
+public interface CarService {
 
-@Service
-public class CarService {
-
-    private final CarRepository carRepository;
-    private CarMapper carMapper;
-
-    public CarService(CarRepository carRepository) {
-        this.carRepository = carRepository;
-    }
-
-    public List<CarResponse> getCars() {
-        return carRepository.findAll()
-                .stream()
-                .map(carMapper::convertToDto)
-                .collect(Collectors.toList());
-    }
+    Page<CarResponse> getCarsPaginated(int page, int size);
 }
